Static distributed tracing (#6)
## What is the goal of this PR?

In grabl tracing, distributed tracing works by creating a trace directly from the client given its parent's `rootId` and `id`. This method was not supported by `GrablTracingThreadStatic` since it was analysis centric, so we have adjusted the API to be client centric so that it can achieve both. This means dependents will need to be updated slightly.

## What are the changes implemented in this PR?

- Drive-by: Added maven coordinates (fixed bazel issues with `java_deps`, but does not include maven deployment)
- Refactored `GrablTracingThreadStatic` to supply both client and analysis level tracing.