Change Maven groupId from io.grabl.tracing to io.grabl (#36)
## What is the goal of this PR?

To allow Grakn to run in Linux.

Currently there is an issue packaging the grabl-tracing-client and grabl-tracing-protocol JARs into the linux targz. This is suspected to be a bug in bazelbuild/rules_pkg (https://github.com/bazelbuild/rules_pkg) but it is triggered by having a file name that is too long. The offending file name is `io-grabl-tracing-grabl-tracing-client-0e4d1618264fbfed67a9b2cbacaed8e4603c1ec7.jar`.

## What are the changes implemented in this PR?

Change Maven groupId from io.grabl.tracing to io.grabl. This should save 8 characters.

We also add `automation.yml` to allow the project to build in Grabl, and ensure every file is covered by a passing `checkstyle_test`. The license type is AGPL.